+++ autodiscovery/pkg/transport/rpc/peerBuilder.go
@@ -16,11 +16,11 @@ func (f PeerBuilder) ToPeerDigest(p discovery.Peer) *api.PeerDigest {
 	return &api.PeerDigest{
 		Identity: &api.PeerIdentity{
 			IP:        p.Identity().IP().String(),
-			Port:      p.Identity().Port(),
+			Port:      uint32(p.Identity().Port()),
 			PublicKey: p.Identity().PublicKey(),
 		},
 		HeartbeatState: &api.PeerHeartbeatState{
-			GenerationTime:    p.HeartbeatState().GenerationTime().Unix(),
+			GenerationTime:    uint64(p.HeartbeatState().GenerationTime().Unix()),
 			ElapsedHeartbeats: p.HeartbeatState().ElapsedHeartbeats(),
 		},
 	}
@@ -29,8 +29,8 @@ func (f PeerBuilder) ToPeerDigest(p discovery.Peer) *api.PeerDigest {
 //FromPeerDigest creates a peer from a digest peer
 func (f PeerBuilder) FromPeerDigest(p *api.PeerDigest) discovery.Peer {
 	return discovery.NewPeerDigest(
-		discovery.NewPeerIdentity(net.ParseIP(p.Identity.IP), p.Identity.Port, p.Identity.PublicKey),
-		discovery.NewPeerHeartbeatState(time.Unix(p.HeartbeatState.GenerationTime, 0), p.HeartbeatState.ElapsedHeartbeats),
+		discovery.NewPeerIdentity(net.ParseIP(p.Identity.IP), uint16(p.Identity.Port), p.Identity.PublicKey),
+		discovery.NewPeerHeartbeatState(time.Unix(int64(p.HeartbeatState.GenerationTime), 0), uint64(p.HeartbeatState.ElapsedHeartbeats)),
 	)
 }
 
@@ -39,11 +39,11 @@ func (f PeerBuilder) ToPeerDiscovered(p discovery.Peer) *api.PeerDiscovered {
 	return &api.PeerDiscovered{
 		Identity: &api.PeerIdentity{
 			IP:        p.Identity().IP().String(),
-			Port:      p.Identity().Port(),
+			Port:      uint32(p.Identity().Port()),
 			PublicKey: p.Identity().PublicKey(),
 		},
 		HeartbeatState: &api.PeerHeartbeatState{
-			GenerationTime:    p.HeartbeatState().GenerationTime().Unix(),
+			GenerationTime:    uint64(p.HeartbeatState().GenerationTime().Unix()),
 			ElapsedHeartbeats: p.HeartbeatState().ElapsedHeartbeats(),
 		},
 		AppState: &api.PeerAppState{
@@ -53,7 +53,7 @@ func (f PeerBuilder) ToPeerDiscovered(p discovery.Peer) *api.PeerDiscovered {
 				Lat: float32(p.AppState().GeoPosition().Lat),
 				Lon: float32(p.AppState().GeoPosition().Lon),
 			},
-			P2PFactor: p.AppState().P2PFactor(),
+			P2PFactor: uint32(p.AppState().P2PFactor()),
 			Status:    api.PeerAppState_PeerStatus(p.AppState().Status()),
 			Version:   p.AppState().Version(),
 		},
@@ -63,8 +63,8 @@ func (f PeerBuilder) ToPeerDiscovered(p discovery.Peer) *api.PeerDiscovered {
 //FromPeerDiscovered creates a peer from a discovered peer
 func (f PeerBuilder) FromPeerDiscovered(p *api.PeerDiscovered) discovery.Peer {
 
-	id := discovery.NewPeerIdentity(net.ParseIP(p.Identity.IP), p.Identity.Port, p.Identity.PublicKey)
-	hb := discovery.NewPeerHeartbeatState(time.Unix(p.HeartbeatState.GenerationTime, 0), p.HeartbeatState.ElapsedHeartbeats)
+	id := discovery.NewPeerIdentity(net.ParseIP(p.Identity.IP), uint16(p.Identity.Port), p.Identity.PublicKey)
+	hb := discovery.NewPeerHeartbeatState(time.Unix(int64(p.HeartbeatState.GenerationTime), 0), p.HeartbeatState.ElapsedHeartbeats)
 	state := discovery.NewPeerAppState(
 		p.AppState.Version,
 		discovery.PeerStatus(p.AppState.Status),
@@ -74,7 +74,7 @@ func (f PeerBuilder) FromPeerDiscovered(p *api.PeerDiscovered) discovery.Peer {
 		},
 		p.AppState.CPULoad,
 		float64(p.AppState.FreeDiskSpace),
-		p.AppState.P2PFactor,
+		uint8(p.AppState.P2PFactor),
 	)
 
 	return discovery.NewDiscoveredPeer(id, hb, state)
